consumidor-mqtt-producer-amqp/pyproject.toml
[tool.poetry]
name = "consumidor-mqtt-producer-amqp"
version = "0.1.0"
description = ""
authors = ["Daniel Alexandre Gouveia <daniel.gouveia@serpro.gov.br>"]

[tool.poetry.dependencies]
python = "^3.8"
paho-mqtt = "^1.6.1"
pika = "^1.2.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
